requirements.txt
PyQt5>=5.15.0
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
lxml>=4.9.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0
pywin32>=306
pythonnet>=3.0.0